separated logic into two tests. The first tests if the mock json input creates a valid PhoneNumberVerification. The second checks the input against what gets placed into the database
